+++ rust/ommx/src/convert/function.rs
@@ -1,0 +1,43 @@
+use crate::v1::{function, function::Function as FunctionEnum, Function, Linear, Quadratic};
+use std::collections::BTreeSet;
+
+impl From<function::Function> for Function {
+    fn from(f: function::Function) -> Self {
+        Self { function: Some(f) }
+    }
+}
+
+impl From<Linear> for Function {
+    fn from(linear: Linear) -> Self {
+        Self {
+            function: Some(function::Function::Linear(linear)),
+        }
+    }
+}
+
+impl From<Quadratic> for Function {
+    fn from(q: Quadratic) -> Self {
+        Self {
+            function: Some(function::Function::Quadratic(q)),
+        }
+    }
+}
+
+impl From<f64> for Function {
+    fn from(f: f64) -> Self {
+        Self {
+            function: Some(function::Function::Constant(f)),
+        }
+    }
+}
+
+impl Function {
+    pub fn used_decision_variable_ids(&self) -> BTreeSet<u64> {
+        match &self.function {
+            Some(FunctionEnum::Linear(linear)) => linear.used_decision_variable_ids(),
+            Some(FunctionEnum::Quadratic(quadratic)) => quadratic.used_decision_variable_ids(),
+            Some(FunctionEnum::Polynomial(poly)) => poly.used_decision_variable_ids(),
+            _ => BTreeSet::new(),
+        }
+    }
+}
